+++ tests/Contract/CocktailContractTest.php
@@ -18,20 +18,23 @@ class CocktailContractTest extends ContractTestCase
         $this->actingAs(
             User::factory()->create()
         );
+
+        $this->setupBar();
     }
 
     public function test_contract_cocktails()
     {
-        Cocktail::factory()->count(10)->create();
+        Cocktail::factory()->count(10)->create(['bar_id' => 1]);
 
-        $response = $this->getJson('/api/cocktails');
+        $response = $this->getJson('/api/cocktails?bar_id=1');
         $response->assertValidResponse(200);
     }
 
     public function test_contract_cocktail_show()
     {
         $cocktail = Cocktail::factory()->create([
-            'name' => 'Test Case'
+            'name' => 'Test Case',
+            'bar_id' => 1
         ]);
 
         $response = $this->getJson('/api/cocktails/' . $cocktail->id);
@@ -46,13 +49,13 @@ class CocktailContractTest extends ContractTestCase
 
     public function test_contract_cocktail_create()
     {
-        $response = $this->postJson('/api/cocktails', [
+        $response = $this->postJson('/api/cocktails?bar_id=1', [
             'name' => "Cocktail name",
             'instructions' => "1. Step\n2. Step"
         ]);
         $response->assertValidRequest()->assertValidResponse(201);
 
-        $response = $this->postJson('/api/cocktails', [
+        $response = $this->postJson('/api/cocktails?bar_id=1', [
             'instructions' => "1. Step\n2. Step"
         ]);
         $response->assertValidRequest()->assertValidResponse(422);
@@ -60,7 +63,7 @@ class CocktailContractTest extends ContractTestCase
 
     public function test_contract_cocktail_update()
     {
-        $cocktail = Cocktail::factory()->create();
+        $cocktail = Cocktail::factory()->create(['bar_id' => 1, 'created_user_id' => 1]);
 
         $response = $this->putJson('/api/cocktails/' . $cocktail->id, [
             'name' => "Cocktail name",
@@ -76,7 +79,7 @@ class CocktailContractTest extends ContractTestCase
 
     public function test_contract_cocktail_delete()
     {
-        $cocktail = Cocktail::factory()->create();
+        $cocktail = Cocktail::factory()->create(['bar_id' => 1, 'created_user_id' => 1]);
 
         $response = $this->deleteJson('/api/cocktails/' . $cocktail->id);
         $response->assertValidResponse(204);